fix api_authenticated pipeline
