+++ client/stylesheets/base/_buttons.scss
@@ -32,4 +32,6 @@ button,
     border-bottom: 0;
     transform: translateY(1px);
   }
+
+  width: 100%;
 }
