+++ lib/transport_web/api/views/discussion_view.ex
@@ -1,15 +1,7 @@
 defmodule TransportWeb.API.DiscussionView do
-  def render(_conn, %{data: data}) do
-    case Poison.encode(data) do
-      {:ok, body} -> body
-      {:error, error} -> error
-    end
-  end
+  alias TransportWeb.API.JSONView
 
-  def render(_conn, %{errors: errors}) do
-    case Poison.encode(errors) do
-      {:ok, body} -> body
-      {:error, error} -> error
-    end
+  def render(conn, assigns) do
+    JSONView.render(conn, assigns)
   end
 end
