+++ priv/gettext/dataset.pot
@@ -54,17 +54,17 @@ msgstr ""
 msgid "Dataset file"
 msgstr ""
 
-#: lib/transport/datagouvfr/client/datasets.ex:25
+#: lib/transport/datagouvfr/client/datasets.ex:26
 #: lib/transport/datagouvfr/client/resources.ex:22
 msgid "A description is needed"
 msgstr ""
 
-#: lib/transport/datagouvfr/client/datasets.ex:45
+#: lib/transport/datagouvfr/client/datasets.ex:46
 #: lib/transport/datagouvfr/client/resources.ex:27
 msgid "A title is needed"
 msgstr ""
 
-#: lib/transport/datagouvfr/client/datasets.ex:40
+#: lib/transport/datagouvfr/client/datasets.ex:41
 msgid "An organization is needed"
 msgstr ""
